requirement.txt
numpy>=1.19.0
scipy>=1.5.0
pandas>=1.1.0
matplotlib>=3.3.0
jupyter>=1.0.0
pytest>=6.0.0
pytest-cov>=2.10.0
qiskit>=0.24.0
pyyaml>=5.3.0